Revert 64bit number conversions added in r19189, don't add the previous long and unsigned long conversions back and use double in the Replay menu.
64bit conversions (including the long ones) are not safe, because not every number can be converted to the 2^53 JS numbers and pretending to do so is asking for bugs.
Explicitly use the double type in the Replay menu, because std::time_t is unspecified and some platforms like Ubuntu yakkety:i386 fail to build, looking for long.
Double should work for the next 285 million years, becomes consistent with SavedGame.cpp, is tested by test_ScriptConversions.cpp and doesn't pretend to cover all 64bit numbers.

Patch By: echotangoecho
Differential Revision: https://code.wildfiregames.com/D205
Refs #3848 D84 D112




git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@19367 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/Replay.cpp
@@ -54,7 +54,7 @@ CReplayLogger::~CReplayLogger()
 void CReplayLogger::StartGame(JS::MutableHandleValue attribs)
 {
 	// Add timestamp, since the file-modification-date can change
-	m_ScriptInterface.SetProperty(attribs, "timestamp", std::time(nullptr));
+	m_ScriptInterface.SetProperty(attribs, "timestamp", (double)std::time(nullptr));
 
 	// Add engine version and currently loaded mods for sanity checks when replaying
 	m_ScriptInterface.SetProperty(attribs, "engine_version", CStr(engine_version));
